+++ src/store/index.ts
@@ -1,12 +1,12 @@
 import { configureStore } from '@reduxjs/toolkit';
-import { reducer } from './reducer.ts';
+import { rootReducer } from './root-reducer.ts';
 import {createApi} from '../services/api/api.ts';
 import {redirect} from './middlewares/redirects.ts';
 
 export const api = createApi();
 
 export const store = configureStore({
-  reducer,
+  reducer: rootReducer,
   middleware: (getDefaultMiddleware) =>
     getDefaultMiddleware({
       thunk: {
